@penelopeysm 2024-12-01
Looping is implemented in reverse now. Had to be careful with indices but it's looking good:

## Accuracy on typical Cholesky samples

![bijector_forward_typical](https://github.com/user-attachments/assets/7b5c7685-81f4-4e7b-ad45-a8dcf75f71a4)

## Performance

```julia
julia> @be (rand(LKJCholesky(5, 1.0, 'U'))) _link_chol_lkj_from_upper_old(_.U)
Benchmark: 4882 samples with 143 evaluations
 min    108.392 ns (2 allocs: 144 bytes)
 median 118.301 ns (2 allocs: 144 bytes)
 mean   125.184 ns (2 allocs: 144 bytes, 0.20% gc time)
 max    12.373 μs (2 allocs: 144 bytes, 97.83% gc time)

julia> @be (rand(LKJCholesky(5, 1.0, 'U'))) _link_chol_lkj_from_upper_new(_.U)
Benchmark: 2924 samples with 241 evaluations
 min    109.959 ns (2 allocs: 144 bytes)
 median 119.295 ns (2 allocs: 144 bytes)
 mean   129.468 ns (2 allocs: 144 bytes, 0.33% gc time)
 max    9.770 μs (2 allocs: 144 bytes, 97.67% gc time)
```

## Unconstrained samples

```julia
julia> y = rand(Random.Xoshiro(468), 10) * 16;

julia> x = _inv_link_chol_lkj_new(y)[1];

julia> y_old = _link_chol_lkj_from_upper_old(x)
ERROR: DomainError with 1.000207932997037:
atanh(x) is only defined for |x| ≤ 1.
Stacktrace:
 [1] atanh_domain_error(x::Float64)
   @ Base.Math ./special/hyperbolic.jl:240
 [2] atanh
   @ ./special/hyperbolic.jl:256 [inlined]
 [3] _link_chol_lkj_from_upper_old(W::Matrix{Float64})
   @ Main ./REPL[60]:12
 [4] top-level scope
   @ REPL[111]:1

julia> y_new = _link_chol_lkj_from_upper_new(x)
10-element Vector{Float64}:
  1.7139942709891685
  4.050190371709019
 12.606351374271206
  8.239542965781226
  7.8978551586304855
  6.885928358454504
  7.201266901997009
  4.588778566499247
  5.507106236959028
 11.582258189742753
```