@devmotion 2024-12-01
Another point: I wonder if generally it would be better to avoid `atanh` completely, due to its constrained domain and its derivative `1/(1 - x^2)` which might be problematic close to 1 and -1. Note that you could rewrite $$\mathrm{atanh}\left(w_{i,j} / \sqrt{\sum_{k=i}^{j} w_{k,j}}\right)$$ as $$\mathrm{asinh}\left(w_{i,j}/\sqrt{\sum_{k={i+1}}^j w_{k,j}}\right)$$.