@penelopeysm 2025-06-19
I finally got the tests to pass, in a rather horrendous way (it's the `test_rrule` change that I'm not fond of; suggestions are welcome, I also posted a question about this on [Slack](https://julialang.slack.com/archives/C6G240ENA/p1750289138078509)). At this point, I'm not sure if we may as well just remove the call to `test_rrule` because it's so lax, or if there's a better way to test it rather than finite differencing.

Also updated all the benchmarks & plots in the top comment.

@devmotion Pinging again in case you want to have another look, but feel free not to if you're busy. Otherwise also cc @sunxd3 for a sanity check -- the top comment contains an explanation of how we arrived at these changes so hopefully that helps in reviewing.